Rename package brooklyn.networking.vclouddirector
Package name was used by two modules (very bad in OSGi) so renaming
to brooklyn.networking.vclouddirector.nat in the bundle
brooklyn-networking-vcloud-director
