+++ vcloud-director/src/test/java/brooklyn/networking/vclouddirector/nat/NatServiceLiveTest.java
@@ -1,4 +1,4 @@
-package brooklyn.networking.vclouddirector;
+package brooklyn.networking.vclouddirector.nat;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 import static org.testng.Assert.assertNotNull;
@@ -8,6 +8,9 @@ import java.util.List;
 import org.testng.annotations.Test;
 
 import com.vmware.vcloud.api.rest.schema.NatRuleType;
+
+import brooklyn.networking.vclouddirector.nat.NatService;
+import brooklyn.networking.vclouddirector.nat.PortForwardingConfig;
 
 import com.google.common.net.HostAndPort;
